Distance Histograms

DistanceHistograms.jl is a lightweight package for optimised computation of pairwise and cross-correlated distance histograms. These summarise distance matrix distributions with constant memory and greater efficiency than compute-then-bin approaches. Inputs are collections of coordinates, given as Vector{SVector}; outputs are Vector{Int64}. All Distances.jl SemiMetrics are supported, with some specialised versions available. Binning is linear and 0->maximum style. This packaged is optimised for medium scale 2-point correlation function computation, and does not employ spatial tree techniques, though the kernels are suitable for this.

Compute an autocorrelation function with auto_corr, or a cross correlation function with cross_corr. Select a suitable maximal value by setting rmax, and the number of bins with nbins. metric may be set to any SemiMetric to modify the distance computation used. blocksize may be used to set the batch size for autocorrelation computation.

We can compute a distance histogram over any set of points,

# generate a set of points
julia> a = collect(CartesianIndices((100,100)));
# convert to SVectors
julia> b = reinterpret(SVector{length(axes(a)),Float64}, Float64.(vec(DistanceHistograms.as_ints(a))));
#computing a correlation function with 10 bins
julia> auto_corr(b,zeros(Int64, 10), 145, 10, metric=Euclidean())
10-element Vector{Int64}:
 2790126
 6994502
 9263358
 9780918
 8944602
 6970206
 3913838
 1113342
  212062
   12046

julia> cross_corr(b,b,zeros(Int64, 10), 145, 10, metric=Euclidean())
10-element Vector{Int64}:
  5590252
 13989004
 18526716
 19561836
 17889204
 13940412
  7827676
  2226684
   424124
    24092

Distance Histograms

Functions

julia> using DistanceHistograms

Functions

DistanceHistograms.auto_corr — Function

auto_corr(points::AbstractVector{T}, histo::Array{Int64,1}, rmax::Int64, nbins::Int64; blocksize::Int=64, metric::Metric=Euclidean())) where {T<:SVector}

Compute the autocorrelation distance histogram for a set of SVector points

Threaded implementation for larger autocorrelations

DistanceHistograms.cross_corr — Function

cross_corr(r1::AbstractVector{T}, r2::AbstractVector{T}, rmax::Int, nbins::Int; metric::Metric=Euclidean()) where {T<:SVector}

Compute the cross correlation distance histogram between a a collection of points r1 and another collection r2

DistanceHistograms.unitary_corr — Function

unitary_corr(r1::AbstractVector{T}, rp, rmax::Int64, nbins::Int64, metric::Metric=Euclidean()) where {T<:SVector}

Compute the cross correlation distance histogram between a single SVector rp and a collection of points r1

DistanceHistograms.auto_corr! — Function

auto_corr!(points::AbstractVector{T}, histo::AbstractVector{Int}, rmax::Int, nbins::Int; blocksize::Int=64, metric::Metric=Euclidean()) where {T<:SVector}

Compute the autocorrelation distance histogram for a set of SVector{T} points assigned to a preallocated array

DistanceHistograms.cross_corr! — Function

cross_corr(r1::AbstractVector{T}, r2::AbstractVector{T}, histo::AbstractVector{Int}, rmax::Int, nbins::Int; metric::Metric=Euclidean()) where {T<:SVector}

Compute the cross correlation distance histogram between a a collection of points r1 and another collection r2 for a preallocated array

DistanceHistograms.unitary_corr! — Function

unitary_corr(r1::AbstractVector{T}, rp, histo, rmax::Int64, nbins::Int64, metric::Metric=Euclidean()) where {T<:SVector}

Compute the cross correlation distance histogram between a single SVector rp and a collection of points r1 and store the result to a preallocated array